feat: add judge regulation is favorited

+++ backend/buptlaw/law-users/src/main/java/edu/bupt/law/user/service/impl/HistoryServiceImpl.java
@@ -77,7 +77,21 @@ public class HistoryServiceImpl implements HistoryService {
 
     @Override
     public Boolean deleteRegulationHistory(String username, Long id){
+        RegulationHistory regulationHistory = regulaitonHistoryRepository.findById(id).orElse(null);
+        if (regulationHistory == null && !regulationHistory.getUser().getUsername().equals(username)) {
+            return false;
+        }
         recordHistoryRepository.deleteById(id);
+        return true;
+    }
+
+    @Override
+    public Boolean isRegulationFavorite(String username, String regulationId) {
+        User user = userRepository.findByUsername(username);
+        RegulationHistory regulationHistory = regulaitonHistoryRepository.findRegulationHistoryByUserAndRegulationIdAndAction(user, regulationId, "02");
+        if (regulationHistory == null) {
+            return false;
+        }
         return true;
     }
 }
